+++ tests/src/Tests/SanitizationsTests.cpp
@@ -124,12 +124,12 @@ fn main()
 {
 	let x: f32 = 0.0;
 	{
-		let i: u32 = u32(0);
-		while (i < (u32(10)))
+		let i: u32 = 0;
+		while (i < 10)
 		{
 			let v: f32 = data.value[i];
 			x += v;
-			i += u32(1);
+			i += 1;
 		}
 
 	}
